backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
python-multipart==0.0.6
